split verify code into separate files
