+++ packages/verifiers/src/index.ts
@@ -1,1 +1,4 @@
+import { getActionSignatureData, getSessionSignatureData } from "./verify_ethereum.js"
+
 export * from "./verify.js"
+export { getActionSignatureData, getSessionSignatureData }
